move tests next into scripts sub dir and refactor code into separate files
