+++ .noble_version_tracker.json
@@ -1,4 +1,4 @@
 {
-  "last_tracked_version": "v11.0.0",
+  "last_tracked_version": "v10.0.0",
   "last_checked": "2025-11-14T14:36:58.293690+00:00"
 }
