+++ scripts/check_noble_version/__init__.py
@@ -1,0 +1,32 @@
+"""
+Noble Documentation Version Tracker
+
+This package provides functionality to track Noble chain versions and
+generate diffs between versions for documentation updates.
+"""
+
+from check_noble_version.version import parse_version, compare_versions
+from check_noble_version.parser import get_latest_version_from_upgrades
+from check_noble_version.tracker import load_tracker, save_tracker
+from check_noble_version.github import get_diff_between_tags, format_diff_summary
+from check_noble_version.config import (
+    REPO_ROOT,
+    MAINNET_MDX_PATH,
+    TRACKER_JSON_PATH,
+    GITHUB_REPO,
+)
+
+__all__ = [
+    "parse_version",
+    "compare_versions",
+    "get_latest_version_from_upgrades",
+    "load_tracker",
+    "save_tracker",
+    "get_diff_between_tags",
+    "format_diff_summary",
+    "REPO_ROOT",
+    "MAINNET_MDX_PATH",
+    "TRACKER_JSON_PATH",
+    "GITHUB_REPO",
+]
+
